Chore: unused packages have been removed

+++ jpa-hib/src/main/java/com/campos/jpa_hib/services/UserService.java
@@ -15,44 +15,46 @@ import java.util.Optional;
 @Service
 public class UserService {
 
-    @Autowired
-    private UserRepository userRepository;
+  @Autowired
+  private UserRepository userRepository;
 
-    public  List<User> findAll() {
-        return userRepository.findAll();
-    }
+  public List<User> findAll() {
+    return userRepository.findAll();
+  }
 
-    public User findById(Long id) {
-        Optional<User> obj = userRepository.findById(id);
-        return obj.orElseThrow(() -> new ResourceNotFoundException(id));
-    }
+  public User findById(Long id) {
+    Optional<User> obj = userRepository.findById(id);
+    return obj.orElseThrow(() -> new ResourceNotFoundException(id));
+  }
 
-    public User create(User user) {
-        return userRepository.save(user);
-    }
+  public User create(User user) {
+    return userRepository.save(user);
+  }
 
-    public User update(Long id, User obj) {
-        try {
-            User entity = userRepository.getReferenceById(id);
-            updateData(entity, obj);
-            return userRepository.save(entity);
-        } catch (EntityNotFoundException e) {
-            throw new ResourceNotFoundException(id);
-        }
+  public User update(Long id, User obj) {
+    try {
+      User entity = userRepository.getReferenceById(id);
+      updateData(entity, obj);
+      return userRepository.save(entity);
+    } catch (EntityNotFoundException e) {
+      throw new ResourceNotFoundException(id);
     }
+  }
 
-    public void updateData(User entity, User obj) {
-        entity.setName(obj.getName());
-        entity.setEmail(obj.getEmail());
-        entity.setPhone(obj.getPhone());
-    }
+  public void updateData(User entity, User obj) {
+    entity.setName(obj.getName());
+    entity.setEmail(obj.getEmail());
+    entity.setPhone(obj.getPhone());
+  }
 
-    public void delete(Long id) {
-        User user = findById(id);
-        try {
-            userRepository.deleteById(id);
-        }  catch (DataIntegrityViolationException e) {
-            throw new DatabaseException(e.getMessage());
-        }
+  public void delete(Long id) {
+    User user = findById(id);
+    if (user != null) {
+      try {
+        userRepository.deleteById(id);
+      } catch (DataIntegrityViolationException e) {
+        throw new DatabaseException(e.getMessage());
+      }
     }
+  }
 }
